feat(mpc-v4): CSDL as the primary Entity reference
This commit makes it that the CSDL that we got from the front-end
is the primary data that we're using.

It may be the case that we need to use complex types for the
modeling. So this also adds an annotation '@segw.expand'.

This annotation reverts back to the old code and way of processing.
This is still kinda broken, so there is a warning message stating
that is experimental. This can be applied on to the Entity Level
or the Entity Property Level.