feat(mpc-v4): Added `@segw.association.fix`
This annotatation will try to fix the warning from the ABAP generator.

SEGW will do a check where if the principal element or the assocation
is nullable the dependent has to be nullable. Vice versa if the
dependent is not nullable, then the principal element and the
association is not nullable.

SAP CAP and the compiler doesn't work with this.

If one creates an entity with an element that is an association,
it would automatically create and element with the key of the
associated entity, but that key is nullable.

One way to get around this is to create assocation and the principal
element manually.

The other way is to automagically fix this for the developer.
Which is not always ideal, because the assocation could be wrong.

With this commit, this is an opt-in way of fixing the association
automagically.